b
b

+++ src/main/java/com/example/SerminaProject/Model/DTO/TourRequestDTO.java
@@ -1,13 +1,15 @@
 package com.example.SerminaProject.Model.DTO;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class TourRequestDTO {
     private String tenTour;
     private String moTa;
-    private Double gia;
+    private BigDecimal gia;
     
     // Hứng cái mảng chứa danh sách các cửa hàng
     private List<TramDTO> danhSachTram;
